+++ tests/testthat/test-rect.R
@@ -11,8 +11,8 @@ test_that("rects equivalent regardless of direction", {
     rect(0.8, 0.8, 0.2, 0.2)
   })
 
-  rect1 <- xml_attrs(xml_find_all(x1, "./rect")[[2]])
-  rect2 <- xml_attrs(xml_find_all(x2, "./rect")[[2]])
+  rect1 <- xml_attrs(xml_find_all(x1, "./g/rect")[[1]])
+  rect2 <- xml_attrs(xml_find_all(x2, "./g/rect")[[1]])
 
   expect_equal(rect1, rect2)
 })
@@ -22,7 +22,7 @@ test_that("fill and stroke colors", {
     plot.new()
     rect(0.2, 0.2, 0.8, 0.8, col = "blue", border = "red")
   })
-  rectangle <- xml_find_all(x, "./rect")[[2]]
+  rectangle <- xml_find_all(x, "./g/rect")[[1]]
   expect_equal(style_attr(rectangle, "fill"), rgb(0, 0, 1))
   expect_equal(style_attr(rectangle, "stroke"), rgb(1, 0, 0))
 })